chore: remove old site files
