+++ package.json
@@ -36,6 +36,12 @@
     "@promethean/piper": "workspace:*",
     "@promethean/pm2-helpers": "workspace:*",
     "@promethean/utils": "workspace:*",
+    "@promethean/health-dashboard-frontend": "workspace:*",
+    "@promethean/llm-chat-frontend": "workspace:*",
+    "@promethean/markdown-graph-frontend": "workspace:*",
+    "@promethean/portfolio-frontend": "workspace:*",
+    "@promethean/smart-chat-frontend": "workspace:*",
+    "@promethean/smartgpt-dashboard-frontend": "workspace:*",
     "@types/node": "24.3.1",
     "@xenova/transformers": "2.17.2",
     "adm-zip": "0.5.16",
