+++ packages/health-dashboard-frontend/static/index.html
