+++ packages/llm-chat-frontend/static/index.html
