+++ packages/markdown-graph-frontend/static/graph.html
