+++ packages/portfolio-frontend/src/frontend/placeholder.js
@@ -1,0 +1,1 @@
+// placeholder
