+++ packages/smart-chat-frontend/static/index.html
