+++ packages/smartgpt-dashboard-frontend/static/index.html
