+++ packages/smartgpt-dashboard-frontend/static/styles.css
