--- sites/portfolio/README.md.bak
@@ -1,8 +1,0 @@
-# Portfolio Site (Placeholder)
-
-This directory is reserved for the personal portfolio previously hosted at [riatzukiza.github.io](https://github.com/riatzukiza/riatzukiza.github.io).
-The source files have not yet been migrated into the Promethean monorepo.
-
-For now the folder only contains this README. References to `npm` scripts from the original project have been removed until the site assets and build configuration are restored.
-
-Progress on migrating the site can be tracked in [docs/MIGRATION_PLAN.md](../docs/MIGRATION_PLAN.md).
